Added static game memory allocation
Removed global sound buffer and game state structs and moved them into
heap-allocated memory

+++ code/common.h
@@ -2,9 +2,9 @@
 #define internal static
 #define local_persist static
 
-#define KILOBYTES(N) N*1024
-#define MEGABYTES(N) N*1024*1024
-#define GIGABYTES(N) N*1024*1024*1024
+#define KILOBYTES(N) ((uint64_t)N*1024L)
+#define MEGABYTES(N) ((uint64_t)N*1024L*1024L)
+#define GIGABYTES(N) ((uint64_t)N*1024L*1024L*1024L)
 
 typedef float real32_t;
 typedef double real64_t;
